typing all modals and Login components
